REFACTOR: add nodemailer config
